Refactor worker CLI options
Move worker CLI optrions into class
Collect CLI-related classes in namespace
Move MetadataSummarizer to model namespace
